Frontend working, still needs to be cleaned up
