+++ apps/monarch/monarch-frontend/src/main.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { BrowserRouter, Link, Route, Routes } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { SearchTherapists } from './app/SearchTherapists';
-import ManageTherapists from './app/ManageTherapists';
-import { ChakraProvider, TabIndicator } from '@chakra-ui/react';
-import { Tabs, TabList, TabPanels, Tab, TabPanel } from '@chakra-ui/react'
+import { ChakraProvider } from '@chakra-ui/react';
+import AdminPage from './app/AdminPage';
 
 const queryClient = new QueryClient();
 
@@ -17,36 +16,12 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
           style={{ display: 'block', maxWidth: 1080, marginInline: 'auto' }}
         >
         <BrowserRouter>
-          <Tabs isManual variant='enclosed'>
-            <TabList>
-              <Tab>
-                <Link to="/">Search</Link>
-              </Tab>
-              <Tab>
-                <Link to="/admin">Manage</Link>
-              </Tab>
-            </TabList>
-            <TabIndicator
-              mt="-1.5px"
-              height="2px"
-              bg="blue.500"
-              borderRadius="1px"
-            />
-            <TabPanels>
-              <TabPanel>
-                <Routes>
-                  <Route path="/" element={<SearchTherapists />} />
-                  <Route path="/admin" element={<ManageTherapists />} />
-                </Routes>
-              </TabPanel>
-            </TabPanels>
-          </Tabs>
-            {/* <Routes>
-              <Route path="/" element={<SearchTherapists />} />
-            </Routes>
-            <Routes>
-              <Route path="/admin" element={<ManageTherapists />} />
-            </Routes> */}
+          <Routes>
+            <Route path="/" element={<SearchTherapists accessToken=''/>} />
+          </Routes>
+          <Routes>
+            <Route path="/admin" element={<AdminPage />} />
+          </Routes>
           </BrowserRouter>
         </main>
       </QueryClientProvider>
